Fix filter with uppercase GPU or provider name
Previously, the values in `QueryFilter.gpu_name`
and `QueryFilter.provider` were assumed to always
be lowercase. However, it was not enforced when
`QueryFilter` was constructed by setting
attributes directly, e.g.
`query_filter.gpu_name = ["A10"]`, or adding
new values to the list, e.g.
`query_filter.gpu_name.append("A10")`. This lead
to incorrect filtering.

Now `QueryFilter.gpu_name` and
`QueryFilter.provider` can have mixed-case values
and it is the responsibility of the reader to
convert them to lowercase for comparisons.