Add public method for GPU memory approximation
The `correct_gpu_memory` method is based on
vastai's `normalize_gpu_memory` but only uses the
known GPU memory if it differs from reported
memory by no more than 7%. This allows to avoid
unintended corrections for vGPUs. The method is
made public so that it can be used in `dstack`.