[Nebius]: Add fabrics list to provider_data (#192)
Associate each Nebius catalog item with a list of
InfiniBand fabrics that it supports. The list of
fabrics and their details is hardcoded until
Nebius exposes it in the API.

Previously, the list of fabrics was hardcoded in
dstack. Moving it to gpuhunt allows to add new
fabrics without a dstack release.